new loops: overrelaxation, staples and sweep checked

+++ include/gaugemonomial.hh
@@ -117,11 +117,11 @@ public:
                   const Float fac = 1.) const override {
     typedef typename accum_type<Group>::type accum;
 #pragma omp parallel for
-    for (size_t x0 = 0; x0 < h.U->getLt(); x0++) {
-      for (size_t x1 = 0; x1 < h.U->getLx(); x1++) {
-        for (size_t x2 = 0; x2 < h.U->getLy(); x2++) {
-          for (size_t x3 = 0; x3 < h.U->getLz(); x3++) {
-            std::vector<size_t> x = {x0, x1, x2, x3};
+    for (int x0 = 0; x0 < h.U->getLt(); x0++) {
+      for (int x1 = 0; x1 < h.U->getLx(); x1++) {
+        for (int x2 = 0; x2 < h.U->getLy(); x2++) {
+          for (int x3 = 0; x3 < h.U->getLz(); x3++) {
+            std::vector<int> x = {x0, x1, x2, x3};
             for (size_t mu = 0; mu < h.U->getndims(); mu++) {
               accum S;
               get_staples_MCMC_step(S, *h.U, x, mu, (*this).xi, (*this).anisotropic);
